Update FaceRecognitionMainFile.py
I have done this commit is related to the Add Student Page. In the previous version of our Add Student Page there were no Face Detection was happening. But now I have modified the code and integrated the Add Student Page with Face Detection and validated the scenario where image was able to get even no face was shown


1. Launch the application
2. Click on Add People
3. Go to Add Student Page
4. Try to Capture the Image without Roll No it should leave an error message 
5. Try to Capture the image without showing face in front of the camera again it should leave an error message 
6. Now capture image by showing your face as well as enter the roll no . It should save the image success will leave an success message .